Updated: TD Flots of AlgoAvancee has been proofread
Fixed minor and major typos within the TD1 (flots) of Algorithmique Avancee to make it cleaner and better to use as a tool to study for the upcoming exam. The other TDs will be proofread as well at a later time.

Also slightly altered some files names to make it more obvious as to what they contain.